requirements.txt
django==1.11
djangorestframework==3.9.4
pandas==0.24.2